# CLN/PERF: remove ndarray.take and platform int conversions