--- a/pandas/algos.pyx
+++ b/pandas/algos.pyx
@@ -1302,7 +1302,7 @@ def group_last_bin_object(ndarray[object, ndim=2] out,
             else:
                 out[i, j] = resx[i, j]
 
-cdef inline float64_t _median_linear(float64_t* a, int n):
+cdef inline float64_t _median_linear(float64_t* a, int n): 
     cdef int i, j, na_count = 0
     cdef float64_t result
     cdef float64_t* tmp