--- a/pandas/core/indexing.py
+++ b/pandas/core/indexing.py
@@ -9,7 +9,8 @@ from pandas.types.common import (is_integer_dtype,
                                  is_categorical_dtype,
                                  is_list_like,
                                  is_sequence,
-                                 is_scalar)
+                                 is_scalar,
+                                 _ensure_int64)
 from pandas.types.missing import isnull, _infer_fill_value
 
 from pandas.core.index import Index, MultiIndex
@@ -1851,23 +1852,11 @@ def maybe_convert_indices(indices, n):
     """ if we have negative indicies, translate to postive here
     if have indicies that are out-of-bounds, raise an IndexError
     """
-    if isinstance(indices, list):
-        indices = np.array(indices)
-        if len(indices) == 0:
-            # If list is empty, np.array will return float and cause indexing
-            # errors.
-            return np.empty(0, dtype=np.int_)
-
-    # cythonize this to one pass?
-    mask = indices < 0
-    if mask.any():
-        # don't mutate original array
-        indices = indices.copy()
-        indices[mask] += n
-    mask = (indices >= n) | (indices < 0)
-    if mask.any():
-        raise IndexError("indices are out-of-bounds")
-    return indices
+    from pandas.algos import take_bounds_check
+    indices = _ensure_int64(indices)
+    out = np.empty(len(indices), dtype='int64')
+    take_bounds_check(indices, out, n)
+    return out
 
 
 def maybe_convert_ix(*args):