--- a/pandas/core/series.py
+++ b/pandas/core/series.py
@@ -1768,7 +1768,6 @@ class Series(base.IndexOpsMixin, strings.StringAccessorMixin,
         elif isinstance(index, MultiIndex):
             from pandas.core.groupby import _lexsort_indexer
             indexer = _lexsort_indexer(index.labels, orders=ascending)
-            indexer = _ensure_platform_int(indexer)
             new_index = index.take(indexer)
         else:
             new_index, indexer = index.sort_values(return_indexer=True,
@@ -2386,9 +2385,8 @@ class Series(base.IndexOpsMixin, strings.StringAccessorMixin,
         if convert:
             indices = maybe_convert_indices(indices, len(self._get_axis(axis)))
 
-        indices = _ensure_platform_int(indices)
         new_index = self.index.take(indices)
-        new_values = self._values.take(indices)
+        new_values = algos.take_nd(self._values, indices, allow_fill=False)
         return self._constructor(new_values,
                                  index=new_index).__finalize__(self)