--- a/pandas/indexes/base.py
+++ b/pandas/indexes/base.py
@@ -1465,37 +1465,37 @@ class Index(IndexOpsMixin, StringAccessorMixin, PandasObject):
     def take(self, indices, axis=0, allow_fill=True,
              fill_value=None, **kwargs):
         nv.validate_take(tuple(), kwargs)
-        indices = _ensure_platform_int(indices)
-        if self._can_hold_na:
+
+        if not self._can_hold_na and allow_fill and fill_value is not None:
+            msg = 'Unable to fill values because {0} cannot contain NA'
+            raise ValueError(msg.format(self.__class__.__name__))
+        else:
             taken = self._assert_take_fillable(self.values, indices,
                                                allow_fill=allow_fill,
                                                fill_value=fill_value,
                                                na_value=self._na_value)
-        else:
-            if allow_fill and fill_value is not None:
-                msg = 'Unable to fill values because {0} cannot contain NA'
-                raise ValueError(msg.format(self.__class__.__name__))
-            taken = self.values.take(indices)
         return self._shallow_copy(taken)
 
     def _assert_take_fillable(self, values, indices, allow_fill=True,
                               fill_value=None, na_value=np.nan):
-        """ Internal method to handle NA filling of take """
-        indices = _ensure_platform_int(indices)
-
-        # only fill if we are passing a non-None fill_value
+        """ internal method to handle NA filling of take """
         if allow_fill and fill_value is not None:
             if (indices < -1).any():
                 msg = ('When allow_fill=True and fill_value is not None, '
                        'all indices must be >= -1')
                 raise ValueError(msg)
-            taken = values.take(indices)
-            mask = indices == -1
-            if mask.any():
-                taken[mask] = na_value
+            else:
+                taken = algos.take_nd(values, indices, allow_fill=allow_fill,
+                                      fill_value=fill_value)
         else:
-            taken = values.take(indices)
+            # provide wraparound semantics if fill_value not specified
+            from pandas.core.indexing import maybe_convert_indices
+            n = values.shape[0]
+            indices = maybe_convert_indices(indices, n)
+            taken = algos.take_nd(values, indices, allow_fill=False)
         return taken
+
+
 
     @cache_readonly
     def _isnan(self):