… pandas/src/algos_take_helper.pxi
… pandas/src/algos_take_helper.pxi.in
… pandas/src/join.pyx
… pandas/tools/merge.py