@chris-b1 2016-08-06
More broadly, is there ever a reason to coerce indexers back to a platform int?  eg. [here](https://github.com/pydata/pandas/blob/master/pandas/core/frame.py#L3198)

Even ignoring the conversion cost, it seems like `take` is faster with the int64s.  (this is Windows 64, python 2.7)

```python
In [1]: np.random.seed(123)

In [2]: a = np.random.randn(1000000)

In [3]: idx = np.random.randint(0, 1000000, size=1000000).astype('int64')

In [4]: idx_plat = idx.astype(np.int_)

In [5]: %timeit a.take(idx)
100 loops, best of 3: 13.9 ms per loop

In [6]: %timeit a.take(idx_plat)
100 loops, best of 3: 17.5 ms per loop
```