@jreback 2016-08-06
there was some discussion of this quite a while ago
it's possible but you still have to work around the problem in Windows 
where take and searchsorted need int32 
but we almost always have int64 indexers

so you could cast as appropriate in only certain situations (vs all of the ensure stuff that we do)

I think we have enough tests that you could try things and see what breaks