@jreback 2016-08-06
pls rebase as just merged: https://github.com/pydata/pandas/pull/13925