@chris-b1 2016-08-06
Ah, so `numpy` on Windows does need a int32 indexer, but only if running 32 bit python.  So I can't just throw away these checks, but I think I can instead make it so int64s are treated like a "platform int" on Win 64 (even though they aren't really).