@chris-b1 2016-08-06
xref #3033 - platform int stuff is a can of worms