@chris-b1 2016-08-08
I guess I'm re-purposing this to now be about platform ints, closing #3033.  Assuming we actually want to do this, I'd still need to adjust a bunch of tests and root out any remaining corner cases, but here's essentially what I've done / think I'm proposing:

   1. use `algos.take_nd` instead of `ndarray.take` throughout
   2. internally, keep all indexers as int64s as much as possible
   3. externally, and at the few numpy interop spots remaining -  cast indexers to `np.intp` -  previously was `np.int_`.  This would techincally be an API change for 64 bit Windows, where those aren't the same.

This generally should help performance on Windows 64, by avoiding unneeded casts and using a better indexer.  I'm hoping neutral, to slighter better (fewer checks) elsewhere.