@chris-b1 2016-08-09
I think I went down the wrong path here - there are places where we were implicitly relying on numpy's boundschecking, and I worry about introducing segfaults with the unchecked `take_nd` - the tests probably caught most everything, but hard to be sure.  

So instead, I think I may just redefine platform_int to `np.intp` (which solves the Windows perf problem) and back out a lot of the other changes.